requirements.txt
Flask==2.2.3
Flask_SocketIO==5.3.4
Flask_Session==0.4.0
Pillow==10.0.0
SQLAlchemy==2.0.10
Werkzeug==2.3.7
mysqlclient==2.1.1
docker==6.1.3
dnspython